+++ Cantera/src/transport/SimpleTransport.h
@@ -335,7 +335,7 @@ namespace Cantera {
      *
      *
      */
-     virtual void getSpeciesFluxes(int ndim, 
+     virtual void getSpeciesFluxes(size_t ndim,
 				  const doublereal* grad_T, 
 				  int ldx, const doublereal* grad_X, 
 				  int ldf, doublereal* fluxes);
@@ -367,7 +367,7 @@ namespace Cantera {
      *                 or greater than the number of species.
      *  @param fluxes  Vector of calculated fluxes
      */
-    virtual void getSpeciesFluxesExt(int ldf, doublereal* fluxes);
+    virtual void getSpeciesFluxesExt(size_t ldf, doublereal* fluxes);
 
   protected:
 
@@ -690,7 +690,7 @@ namespace Cantera {
     /*!
      * Either 1, 2, or 3
      */
-    int m_nDim;
+    size_t m_nDim;
 
   private:
